Rename project to QuickControllers.
