+++ QuickControllers/App.xaml
@@ -1,8 +1,8 @@
 ﻿<Application
-    x:Class="UWPRobotController.App"
+    x:Class="QuickControllers.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:UWPRobotController"
+    xmlns:local="using:QuickControllers"
     RequestedTheme="Dark">
 
 </Application>
