+++ QuickControllers/ArduinoConnectionPage.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="UWPRobotController.ArduinoConnectionPage"
+    x:Class="QuickControllers.ArduinoConnectionPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:UWPRobotController"
