+++ QuickControllers/Assets/SmallLogo.scale-100.png
