+++ QuickControllers/Assets/SplashScreen.scale-100.png
