+++ QuickControllers/Control/IControl.cs
@@ -1,4 +1,4 @@
-﻿namespace UWPRobotController
+﻿namespace QuickControllers 
 {
     interface IControl
     {
