+++ QuickControllers/Interfaces/DpadControlPage.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="UWPRobotController.Interfaces.DpadControlPage"
+    x:Class="QuickControllers.Interfaces.DpadControlPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:UWPRobotController.Interfaces"
