+++ QuickControllers/Interfaces/DpadControlPage.xaml.cs
@@ -15,7 +15,7 @@ using Windows.UI.Xaml.Navigation;
 
 // The Blank Page item template is documented at http://go.microsoft.com/fwlink/?LinkId=234238
 
-namespace UWPRobotController.Interfaces
+namespace QuickControllers.Interfaces
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
