+++ QuickControllers/Interfaces/TankControlPage.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="UWPRobotController.Interfaces.TankControlPage"
+    x:Class="QuickControllers.Interfaces.TankControlPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:UWPRobotController"
